Add Support To Create Kamaji CKP Cluster in External Target Cluster
This patch adds feature to deploy tenant control plane in a separate
target cluster rather than using the current cluster(which is also our
CAPI management cluster).

The patch introduces env `EXTERNAL_CLUSTER_KUBECONFIG_SECRET` in the
compass-cluster-manager pod for this.

When `clusterManager.externalClusterReference.kubeconfigSecretName` is
set to a secret name present in compass-capi namespace. Then that secret
will be used to connect and provision tenant control plane(TCP) pods in
that target cluster. If kubeconfigSecretName is not set, i.e. "",(which
is the default behaviour) then current cluster will be used to deploy
the TCP pods.

Note: The secret name provided in `kubeconfigSecretName` should be
placed in namespace `compass-capi` which is default namespace
for CAPI components. The kubeconfig file should be base64 encoded string
with key name `kubeconfig`, i.e:

```yaml
apiVersion: v1
kind: Secret
metadata:
  name: external-cluster-kubeconfig
  namespace: compass-capi
data:
  kubeconfig: <base64 encoded target kubeconfig>
```

The name `external-cluster-kubeconfig` then needs to be passed as value for
`clusterManager.externalClusterReference.kubeconfigSecretName`.

Jira ID: FLUID-464

Signed-off-by: Parth Yadav <parth@coredge.io>